Updated user saving, added applicationuser directly

+++ backend/Controllers/UsersController.cs
@@ -10,6 +10,8 @@ using backend.Data;
 using backend.Models;
 using backend.Types;
 using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+using Microsoft.AspNetCore.Server.Kestrel.Transport.Libuv.Internal.Networking;
 
 namespace backend.Controllers
 {
@@ -18,7 +20,9 @@ namespace backend.Controllers
         private readonly ApplicationDbContext _context;
 
         private readonly UserManager<ApplicationUser> _userManager;
+
         private async Task<ApplicationUser> GetUser() => await _userManager.FindByNameAsync(User.Identity.Name);
+
         // Does not work with a variable, need to be a method
         private Role GetRole() => GetUser().Result.User.Role;
 
@@ -60,18 +64,12 @@ namespace backend.Controllers
         // GET: Users/Create
         public IActionResult Create()
         {
-            if (GetRole() == Role.ReadOnly || GetRole() == Role.Elevated)
+            if (GetRole() != Role.Admin)
             {
-                return NotFound();
+                return RedirectToAction("Index", "Home");
             }
 
-            ViewBag.ApplicationUsers = new MultiSelectList(
-                _context.ApplicationUsers,
-                "Id",
-                "Email"
-            );
-
-            return View();
+            return View(new UserView());
         }
 
         // POST: Users/Create
@@ -79,60 +77,121 @@ namespace backend.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,FirstName,Infix,LastName,Locale,Type,Role")]
-            User user)
+        public IActionResult Create(UserView newUser)
         {
+            if (GetRole() != Role.Admin)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (ModelState.IsValid)
             {
-                _context.Add(user);
-                await _context.SaveChangesAsync();
+                _context.Users.Add(newUser.User);
+                _context.SaveChanges();
+
+                // Create new applicationuser
+                var appUser = new ApplicationUser
+                {
+                    Email = newUser.Email,
+                    NormalizedEmail = newUser.Email.ToUpper(),
+                    UserName = newUser.Email,
+                    NormalizedUserName = newUser.Email.ToUpper(),
+                    PhoneNumber = newUser.Phone,
+                    EmailConfirmed = true,
+                    PhoneNumberConfirmed = true,
+                    SecurityStamp = Guid.NewGuid().ToString("D"),
+                    User = newUser.User
+                };
+
+                var password = new PasswordHasher<ApplicationUser>();
+                var hashed = password.HashPassword(appUser, newUser.Password);
+                appUser.PasswordHash = hashed;
+
+                var userStore = new UserStore<ApplicationUser>(_context);
+                var result = userStore.CreateAsync(appUser);
+
+                // Wait for the actual result
+                result.Wait();
+
                 return RedirectToAction("Index", nameof(People));
             }
 
-            return View(user);
+            return View(newUser);
         }
 
         // GET: Users/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (GetRole() != Role.Admin)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (id == null)
             {
                 return NotFound();
             }
 
             var user = await _context.Users.SingleOrDefaultAsync(m => m.Id == id);
-            if (user == null)
+            var appUser = _context.ApplicationUsers.FirstOrDefault(u => u.UserID == id);
+
+            if (user == null || appUser == null)
             {
                 return NotFound();
             }
 
-            return View(user);
+            var uv = new UserView
+            {
+                User = user,
+                Email = appUser.Email,
+                Phone = appUser.PhoneNumber
+            };
+
+            return View(uv);
         }
 
-        // POST: Users/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        /**
+         * POST for user edit screen
+         */
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,FirstName,Infix,LastName,Locale,Type,Role,Deleted")]
-            User user)
+        public IActionResult Edit(int id, UserView updatedUser)
         {
-            if (id != user.Id)
+            if (GetRole() != Role.Admin)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            if (id != updatedUser.User.Id)
             {
                 return NotFound();
             }
-
 
             if (ModelState.IsValid)
             {
                 try
                 {
-                    _context.Update(user);
-                    await _context.SaveChangesAsync();
+                    _context.Update(updatedUser.User);
+                    _context.SaveChanges();
+
+                    var appUser = _context.ApplicationUsers.FirstOrDefault(u => u.UserID == updatedUser.User.Id);
+                    if (appUser != null)
+                    {
+                        appUser.Email = updatedUser.Email;
+                        appUser.NormalizedEmail = updatedUser.Email;
+                        appUser.PhoneNumber = updatedUser.Phone;
+
+                        _context.ApplicationUsers.Update(appUser);
+                        _context.SaveChanges();
+                    }
+                    else
+                    {
+                        Console.WriteLine("ApplicationUser could not be found.");
+                    }
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (Exception)
                 {
-                    if (!UserExists(user.Id))
+                    if (!UserExists(updatedUser.User.Id))
                     {
                         return NotFound();
                     }
@@ -145,11 +204,21 @@ namespace backend.Controllers
                 return RedirectToAction("Index", nameof(People));
             }
 
-            return View(user);
+
+            return View(new UserView());
         }
 
+
+        /**
+         * Remove functionality
+         */
         public async Task<IActionResult> Delete(int? id)
         {
+            if (GetRole() != Role.Admin)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -168,6 +237,11 @@ namespace backend.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (GetRole() != Role.Admin)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             var @user = await _context.Users.SingleOrDefaultAsync(m => m.Id == id);
             _context.Users.Remove(@user);
             await _context.SaveChangesAsync();
